refactor(expression-runtime): Extract shared serialize module, add proxy detection
